STEP-28 Implemented Authentication part like while login otp and adding members and diff roles

+++ syllabus-gen/src/components/common/RoleUi.jsx
@@ -23,9 +23,12 @@ export default function RolePopup({role,setRole,closePopup}) {
   };
 
   const handleClose = () => {
-    setRole("none");
-    localStorage.setItem("role", "none");
-    closePopup();
+    if(confirm("Are u sure want to view web as a guest?")){
+      setRole("none");
+      localStorage.setItem("role", "none");
+      closePopup();
+      nav('/home')
+    }
   };
 
   return (
